Refactor: extract core runtime traits to shared files in core module
