+++ core/src/main/scala/org/mockito/IdiomaticVerifications.scala
@@ -1,0 +1,9 @@
+package org.mockito
+
+trait IdiomaticVerifications {
+
+  type Verification
+
+  def verification(v: => Any): Verification
+
+}
